# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "d43ecdfd17aca209ecd9b0a8b370e2c0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register a user.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "be2e78e5f93182d4403f839cc516d312",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email address",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login a user.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logs out current logged in user session",
                "description": "",
                "operationId": "logoutUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Success with some route data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/capsules": {
            "get": {
                "tags": [
                    "capsule"
                ],
                "operationId": "514c24b1989be61b583c526782c960f7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of capsules.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/capsules/{capsule_serial}": {
            "get": {
                "tags": [
                    "capsule"
                ],
                "operationId": "d8b60627dd0018799b6c9b5266b08ad3",
                "parameters": [
                    {
                        "name": "capsule_serial",
                        "in": "path",
                        "description": "Capsule Serial",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a capsule.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resource.json": {
            "get": {
                "responses": {
                    "200": {
                        "description": "An example resource"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}